svirt: Fix serial backend for s390x
s390x needs to use <console> for another serial device instead of
<serial>. On the other hand it looks like x86 needs <serial>.
Therefore new parameter is pty_dev is added to consoles::sshVirtsh
and used in start_serial_grab().

Although this backend is probably not going to be used on x86 (plain
qemu backend with virtio console is better option), keep <serial> as
default. This backend is planned to be used on s390x.

+ reuse add_pty() in add_serial_console() as <alias> tag is not needed.

Fixes: 9b618c7c ("Allow adding additional serial port when using svirt backend")